LayoutTests/imported/w3c/web-platform-tests/svg/types/scripted/SVGGraphicsElement.getBBox-03-expected.txt


PASS rect1
PASS rect2
PASS circle
FAIL ellipse1 assert_equals: ellipse1: x expected 0 but got -9
FAIL ellipse2 assert_equals: ellipse2: x expected 0 but got 1
FAIL image3 assert_equals: image3: x expected 0 but got 1
FAIL image4 assert_equals: image4: x expected 0 but got 1
FAIL foreign1 assert_equals: foreign1: width expected 0 but got -10
FAIL foreign2 assert_equals: foreign2: width expected 0 but got 10
PASS SVGGraphicsElement.prototype.getBBox for elements with negative sizes
